better permissions management concerning projects and project_identifiers

+++ application/views/identifiers/show.php
@@ -6,7 +6,9 @@ ob_start();
 <h1 class="mt-5">
   <?php echo $project->name; ?>
   <a class="btn btn-outline-primary" href="/project/<?php echo $project->id; ?>" role="button">Voir le projet</a>
-  <a class="btn btn-outline-primary" href="/identifiers/<?php echo $project->id; ?>/new" role="button">Ajouter un identifiant</a>
+  <?php if ($isMyProject || $controller->hasPermission('project_identifiers', 'add')): ?>
+    <a class="btn btn-outline-primary" href="/identifiers/<?php echo $project->id; ?>/new" role="button">Ajouter un identifiant</a>
+  <?php endif; ?>
 </h1>
 <p class="lead">Les différents identifiants pour le projet</p>
 
@@ -26,9 +28,15 @@ ob_start();
       <td><?php echo nl2br($identifier->value); ?></td>
       <td><?php echo ($identifier->confidential == 1) ? 'Oui' : 'Non'; ?></td>
       <td>
-        <a href="/identifiers/project_edit/<?php echo $identifier->id; ?>">Modifier</a>
-        -
-        <a href="/identifiers/project_delete/<?php echo $identifier->id; ?>">Supprimer</a>
+        <?php if ($isMyProject || $controller->hasPermission('project_identifiers', 'edit')): ?>
+          <a href="/identifiers/project_edit/<?php echo $identifier->id; ?>">Modifier</a>
+        <?php endif; ?>
+        <?php if ($isMyProject || $controller->hasPermission('project_identifiers', 'edit') || $controller->hasPermission('project_identifiers', 'delete')): ?>
+          -
+        <?php endif; ?>
+        <?php if ($isMyProject || $controller->hasPermission('project_identifiers', 'delete')): ?>
+          <a href="/identifiers/project_delete/<?php echo $identifier->id; ?>">Supprimer</a>
+        <?php endif; ?>
       </td>
     </tr>
     <?php endforeach; ?>
