+++ application/views/projects/show.php
@@ -5,9 +5,15 @@ ob_start();
 
 <h1 class="mt-5">
   <?php echo $project->name; ?>
-  <a class="btn btn-outline-primary" href="/identifiers/show/<?php echo $project->id; ?>" role="button">Identifiants</a>
-  <a class="btn btn-outline-primary" href="/project/edit/<?php echo $project->id; ?>" role="button">Modifier</a>
-  <a class="btn btn-outline-danger" href="/project/delete/<?php echo $project->id; ?>" role="button">Supprimer</a>
+  <?php if ($isMyProject || $controller->hasPermission('project_identifiers', 'show')): ?>
+    <a class="btn btn-outline-primary" href="/identifiers/show/<?php echo $project->id; ?>" role="button">Identifiants</a>
+  <?php endif; ?>
+  <?php if ($isMyProject || $controller->hasPermission('projects', 'edit')): ?>
+    <a class="btn btn-outline-primary" href="/project/edit/<?php echo $project->id; ?>" role="button">Modifier</a>
+  <?php endif; ?>
+  <?php if ($controller->hasPermission('projects', 'delete')): ?>
+    <a class="btn btn-outline-danger" href="/project/delete/<?php echo $project->id; ?>" role="button">Supprimer</a>
+  <?php endif; ?>
 </h1>
 <p class="lead">Affichages d'informations concernant le projet</p>
 <p>
